// types/dto.ts
export interface UrlInfo {
  url: string;
  type: "Link" | "Youtube" | "Profile" | "Homepage" | "Blog";
}
// ...
export interface UserDto {
  /** @format int64 */
  id: number;
  name: string;
  bookIsbnList?: number[];
}
// ...
export interface ContentsDto {
  /** @format int64 */
  id: number;
  urls: UrlInfo[];
  title?: string;
  image?: string;
  creator: UserDto;
}